Sending angle deviation from PC Preparing calculation of state scaled to physical units on chip Removing unnecessary angle hanging variable

+++ Firmware/CartPole CubeIDE/Src/communication_with_PC.h
@@ -43,7 +43,7 @@ void prepare_message_to_PC_control_config(
 		unsigned short control_period,
 		bool controlSync,
 		int controlLatencyUs,
-		short angle_setPoint,
+		float angle_setPoint,
 		unsigned short angle_averageLen);
 
 void prepare_message_to_PC_config_PID(
